+++ backend/src/main/java/com/easydiet/backend/dto/FoodResponse.java
@@ -12,7 +12,7 @@ public class FoodResponse {
 
     private UUID id;
     private String name;
-    private Integer calories;
+    private BigDecimal calories;
     private BigDecimal protein;
     private BigDecimal carbs;
     private BigDecimal fat;
